Changes Window.RequestAnimationFrame return type to 'long' (missing save). [#247]

+++ Html5/Window.cs
@@ -671,7 +671,7 @@ namespace Bridge.Html5
         /// </summary>
         /// <param name="callback">A parameter specifying a function to call when it's time to update your animation for the next repaint. The callback has one single argument, a DOMHighResTimeStamp, which indicates the current time for when requestAnimationFrame starts to fire callbacks.</param>
         /// <returns>requestID is a long integer value that uniquely identifies the entry in the callback list. This is a non-zero value, but you may not make any other assumptions about its value. You can pass this value to window.cancelAnimationFrame() to cancel the refresh callback request.</returns>
-        public static int RequestAnimationFrame(Action callback)
+        public static long RequestAnimationFrame(Action callback)
         {
             return default(long);
         }
